requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
pandas>=2.2.0
matplotlib>=3.8.0
